+++ utils/fileSystem/index.web.ts
@@ -7,3 +7,14 @@ export const moveAssetsForMessage = async (
   fromMessageId: string,
   toMessageId: string
 ) => {};
+
+export const createTemporaryDirectory = async () => {};
+
+export const copyDatabasesToTemporaryDirectory = async (
+  tempDirectory: string
+) => {};
+
+export const moveTemporaryDatabasesToDatabaseDirecory = async (
+  tempDirectory: string,
+  inboxId: string
+) => {};
